add mario.config for loading bar and console logs and improve app.js

+++ js/app.js
@@ -2,16 +2,20 @@
 
 let mario = angular.module('app', ['leaflet-directive', 'angular-loading-bar'])
 
+mario.config(['cfpLoadingBarProvider', '$logProvider', 'configProvider', function (cfpLoadingBarProvider, $logProvider, config) {
+  cfpLoadingBarProvider.includeSpinner = false
+  $logProvider.debugEnabled(false)
+}])
+
 mario.controller('Controller',
   ['$scope', 'modifyMap', 'poiService', 'parkingService', 'directRouteService', 'interRouteService', 'config',
   function ($scope, modifyMap, poiService, parkingService, directRouteService, interRouteService, config) {
     let initialize = function () {
-      angular.extend($scope, config)
-      $scope.model.map.layers.baselayers = $scope.model.map.baselayers
+      Object.assign($scope, config)
+      $scope.model.map.layers['baselayers'] = $scope.model.map.baselayers
       directRouteService.getInitialInformation($scope.model)
       interRouteService.modelDate($scope.model, new Date())
     }
-
     initialize()
 
     $scope.changeBaselayer = function (layer) {
@@ -23,17 +27,15 @@ mario.controller('Controller',
     }
 
     $scope.selectPoi = function (category) {
-      category === 'parking'
-      ? parkingService.selectParking($scope.model)
-      : poiService.selectPoi($scope.model, category)
+      poiService.selectPoi($scope.model, category)
     }
 
     $scope.calculate = function () {
-      if ($scope.model.map.markers.length > 1) directRouteService.calculateRoute($scope.model)
+      if ($scope.model.map.markers.length === 2) directRouteService.calculateRoute($scope.model)
     }
 
     $scope.calculateIntermodal = function (range) {
-      if ($scope.model.map.markers.length > 1) interRouteService.calculateIntermodal($scope.model, range)
+      if ($scope.model.map.markers.length === 2) interRouteService.calculateIntermodal($scope.model, range)
     }
 
     $scope.selectAlgo = function (algorithm) {
@@ -62,7 +64,7 @@ mario.controller('Controller',
     }
 
     $scope.hoverSegment = function (index, flag) {
-      $scope.model.selected.hover = flag ? index : -1
+      $scope.model.selected['hover'] = flag ? index : -1
       modifyMap.highlightSegment($scope.model, index, flag)
     }
 
@@ -79,15 +81,11 @@ mario.controller('Controller',
     })
 
     $scope.$on('leafletDirectiveGeoJson.mouseout', function (event, args) {
-      if (args.target) {
-        modifyMap.handleMousOutGeoJson($scope.model, event, args)
-      }
+      if (args.target) modifyMap.handleMousOutGeoJson($scope.model, event, args)
     })
 
     $scope.$watch('model.map.markers[1]', function (newValue, oldValue) {
-      if (oldValue !== newValue && $scope.model.map.markers[1]) {
-        poiService.removePoi($scope.model)
-      }
+      if (oldValue !== newValue && $scope.model.map.markers[1]) poiService.removePoi($scope.model)
     })
   }]
 )
